+++ src/Functions/multiFuzzyMatchAnyIndex.cpp
@@ -1,0 +1,24 @@
+#include "FunctionsMultiStringFuzzySearch.h"
+#include "FunctionFactory.h"
+#include "MultiMatchAnyImpl.h"
+
+
+namespace DB
+{
+
+struct NameMultiFuzzyMatchAnyIndex
+{
+    static constexpr auto name = "multiFuzzyMatchAnyIndex";
+};
+
+using FunctionMultiFuzzyMatchAnyIndex = FunctionsMultiStringFuzzySearch<
+    MultiMatchAnyImpl<UInt64, false, true, true>,
+    NameMultiFuzzyMatchAnyIndex,
+    std::numeric_limits<UInt32>::max()>;
+
+void registerFunctionMultiFuzzyMatchAnyIndex(FunctionFactory & factory)
+{
+    factory.registerFunction<FunctionMultiFuzzyMatchAnyIndex>();
+}
+
+}
