+++ src/Functions/replaceRegexpAll.cpp
@@ -1,0 +1,21 @@
+#include "FunctionStringReplace.h"
+#include "FunctionFactory.h"
+#include "ReplaceRegexpImpl.h"
+
+
+namespace DB
+{
+
+struct NameReplaceRegexpAll
+{
+    static constexpr auto name = "replaceRegexpAll";
+};
+
+using FunctionReplaceRegexpAll = FunctionStringReplace<ReplaceRegexpImpl<false>, NameReplaceRegexpAll>;
+
+void registerFunctionReplaceRegexpAll(FunctionFactory & factory)
+{
+    factory.registerFunction<FunctionReplaceRegexpAll>();
+}
+
+}
