+++ src/Functions/replaceRegexpOne.cpp
@@ -1,0 +1,21 @@
+#include "FunctionStringReplace.h"
+#include "FunctionFactory.h"
+#include "ReplaceRegexpImpl.h"
+
+
+namespace DB
+{
+
+struct NameReplaceRegexpOne
+{
+    static constexpr auto name = "replaceRegexpOne";
+};
+
+using FunctionReplaceRegexpOne = FunctionStringReplace<ReplaceRegexpImpl<true>, NameReplaceRegexpOne>;
+
+void registerFunctionReplaceRegexpOne(FunctionFactory & factory)
+{
+    factory.registerFunction<FunctionReplaceRegexpOne>();
+}
+
+}
